Move topic-related features/steps in the right place
