+++ functional_tests/features/statistical_article_page.feature
@@ -106,56 +106,3 @@ Feature: Statistical Article Page components
         And the user clicks "Publish"
         And the user clicks "View Live" on the publish confirmation banner
         Then the published statistical article page has the added headline figures
-
-    Scenario: A CMS user can choose headline figures when editing a topic page
-        When the user goes to add a new statistical article page
-        And the user adds basic statistical article page content
-        And the user adds headline figures
-        And the user clicks "Publish"
-        And the user edits the ancestor topic
-        And the user clicks to add headline figures to the topic page
-        Then the headline figures are shown
-
-    Scenario: A CMS user can add headline figures to a topic page
-        When the user goes to add a new statistical article page
-        And the user adds basic statistical article page content
-        And the user adds headline figures
-        And the user clicks "Publish"
-        And the user edits the ancestor topic
-        And the user adds two headline figures to the topic page
-        And the user clicks "Publish"
-        And the user clicks "View Live" on the publish confirmation banner
-        Then the published topic page has the added headline figures
-        And the headline figures on the topic page link to the statistical page
-
-    Scenario: A CMS user can reorder headline figures on a topic page
-        When the user goes to add a new statistical article page
-        And the user adds basic statistical article page content
-        And the user adds headline figures
-        And the user clicks "Publish"
-        And the user edits the ancestor topic
-        And the user adds two headline figures to the topic page
-        And the user clicks "Publish"
-        And the user clicks "View Live" on the publish confirmation banner
-        Then the published topic page has the added headline figures in the correct order
-        And the user edits the ancestor topic
-        When the user reorders the headline figures on the topic page
-        And the user clicks "Publish"
-        And the user clicks "View Live" on the publish confirmation banner
-        Then the published topic page has reordered headline figures
-
-    Scenario: A CMS user can reorder headline figures on a Statistical Article Page without affecting the order of the figures on the topic page
-        When the user goes to add a new statistical article page
-        And the user adds basic statistical article page content
-        And the user adds headline figures
-        And the user clicks "Publish"
-        And the user edits the ancestor topic
-        And the user adds two headline figures to the topic page
-        And the user clicks "Publish"
-        And the user clicks "View Live" on the publish confirmation banner
-        Then the published topic page has the added headline figures in the correct order
-        When the user returns to editing the statistical article page
-        And the user reorders the headline figures on the Statistical Article Page
-        And the user clicks "Publish"
-        And the user views the topic page
-        Then the published topic page has the added headline figures in the correct order
