+++ functional_tests/steps/statistical_article_page.py
@@ -153,43 +153,6 @@ def user_adds_headline_figures(context: Context):
     panel.get_by_label("Supporting text*").nth(1).fill("Second supporting text")
 
 
-@step("the user edits the ancestor topic")
-def user_edits_the_ancestor_topic(context: Context):
-    edit_url = reverse("wagtailadmin_pages:edit", args=(context.topic_page.id,))
-    context.page.goto(f"{context.base_url}{edit_url}")
-
-
-@step("the user views the topic page")
-def user_views_the_topic_page(context: Context):
-    context.page.goto(f"{context.base_url}{context.topic_page.url}")
-
-
-@step("the user clicks to add headline figures to the topic page")
-def user_clicks_to_add_headline_figures_to_the_topic_page(context: Context, *, button_index: int = 0):
-    page = context.page
-    panel = page.locator("#panel-child-content-headline_figures-content")
-    panel.get_by_role("button", name="Insert a block").nth(button_index).click()
-    page.wait_for_timeout(100)
-    panel.get_by_role("button", name="Choose Article Series page and headline figure").click()
-    page.wait_for_timeout(100)  # Wait for modal to open
-
-
-@step("the user adds two headline figures to the topic page")
-def user_adds_two_headline_figures_to_the_topic_page(context: Context):
-    page = context.page
-    user_clicks_to_add_headline_figures_to_the_topic_page(context)
-    page.locator(".modal-content").get_by_role("link", name="PSF").nth(0).click()
-    user_clicks_to_add_headline_figures_to_the_topic_page(context, button_index=1)
-    page.locator(".modal-content").get_by_role("link", name="PSF").nth(1).click()
-
-
-@step("the user reorders the headline figures on the topic page")
-def user_reorders_the_headline_figures_on_the_topic_page(context: Context):
-    page = context.page
-    panel = page.locator("#panel-child-content-headline_figures-content")
-    panel.get_by_role("button", name="Move up").nth(1).click()
-
-
 @step("the user reorders the headline figures on the Statistical Article Page")
 def user_reorders_the_headline_figures_on_the_statistical_article_page(context: Context):
     page = context.page
@@ -277,32 +240,6 @@ def the_published_statistical_article_page_has_the_added_headline_figures(contex
     expect(page.get_by_text("~321%")).to_be_visible()
     expect(page.get_by_text("First supporting text")).to_be_visible()
     expect(page.get_by_text("Second supporting text")).to_be_visible()
-
-
-@then("the published topic page has the added headline figures in the correct order")
-def the_published_topic_page_has_the_added_headline_figures_in_the_correct_order(context: Context):
-    page = context.page
-    headline_block = page.locator(".headline-figures .ons-grid__col")
-    expect(headline_block.nth(0).get_by_text("First headline figure")).to_be_visible()
-    expect(headline_block.nth(1).get_by_text("Second headline figure")).to_be_visible()
-
-
-@then("the published topic page has reordered headline figures")
-def the_published_topic_page_has_reordered_headline_figures(context: Context):
-    page = context.page
-    headline_block = page.locator(".headline-figures .ons-grid__col")
-    expect(headline_block.nth(0).get_by_text("Second headline figure")).to_be_visible()
-    expect(headline_block.nth(1).get_by_text("First headline figure")).to_be_visible()
-
-
-@then("the headline figures on the topic page link to the statistical page")
-def the_headline_figures_on_the_topic_page_link_to_the_statistical_page(context: Context):
-    page = context.page
-    page.get_by_text("First headline figure").click()
-    expect(page.get_by_role("heading", name="The article page")).to_be_visible()
-    page.go_back()
-    page.get_by_text("Second headline figure").click()
-    expect(page.get_by_role("heading", name="The article page")).to_be_visible()
 
 
 @then('the user can click on "Show detail" to expand the corrections and notices block')
